requirements.txt
Flask==2.2.2
PyJWT==2.6.0
waitress==2.1.2
